fix(template): rename internal/logger to log
